+++ bootstrap.py
@@ -27,21 +27,17 @@ Author: tjado <https://github.com/tejado>
         TC    <reddit.com/u/Tr4sHCr4fT>
 """
 
-import os
-import sys
-import time
-import json
-import argparse
+import os, time, json
+import argparse, logging
 import sqlite3
-import logging
 
 from s2sphere import CellId, LatLng
 from pgoapi.exceptions import NotLoggedInException
-from fastmap.utils import api_init, get_response, susub_cells, get_cell_ids
-from fastmap.utils import check_db, init_db, set_bit, cover_circle, cover_square
+from fastmap.db import check_db, fill_db
+from fastmap.apiwrap import api_init, get_response
+from fastmap.utils import set_bit, get_cell_ids, susub_cells, cover_circle, cover_square
 
 log = logging.getLogger(__name__)
-sys.path.append(os.path.dirname(os.path.realpath(__file__)))
 
 def init_config():
     parser = argparse.ArgumentParser()     
@@ -96,7 +92,7 @@ def init_config():
         elif config.width:
             cells = cover_square(lat, lng, config.width, config.level)
         else: log.error('Area size not given!'); return
-        log.info('Added %d cells to scan queue.' % init_db(cells, config.dbfile))
+        log.info('Added %d cells to scan queue.' % fill_db(config.dbfile, cells))
         del cells, lat, lng
     
     return config
